backend/requirements.txt
aioboto3==13.2.0
aiobotocore==2.15.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aioitertools==0.12.0
aiosignal==1.3.1
aiosmtplib==3.0.2
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncmy==0.2.9
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.2.0
black==24.10.0
boto3==1.35.36
botocore==1.35.36
botocore-stubs==1.35.70
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.16.1
flake8==7.1.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
jmespath==1.0.1
loguru==0.7.2
Mako==1.3.5
MarkupSafe==3.0.1
mccabe==0.7.0
multidict==6.1.0
mypy==1.12.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pip-system-certs==4.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
propcache==0.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyflakes==3.2.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0
pytest_async==0.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.18
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
stripe==11.1.1
tomli==2.0.2
types-aioboto3==13.2.0.post1
types-aiobotocore==2.15.2.post1
types-aiobotocore-cloudformation==2.15.2
types-aiobotocore-dynamodb==2.15.2
types-aiobotocore-ec2==2.15.2
types-aiobotocore-lambda==2.15.2
types-aiobotocore-rds==2.15.2
types-aiobotocore-s3==2.15.2
types-aiobotocore-sqs==2.15.2
types-awscrt==0.23.1
types-passlib==1.7.7.20240819
types-pyasn1==0.6.0.20240913
types-python-jose==3.3.4.20240106
types-s3transfer==0.10.4
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
virtualenv==20.27.0
wrapt==1.16.0
yarl==1.18.0
